feat: lightweight GET API for campaigns (#208)
A refactor of the endpoint that determines whether a Campaign should be shown as a lightweight, standalone application that uses WordPress database and cacheing libraries but doesn't incur the full weight of a WordPress REST API request. On a site with campaigns this endpoint will often be hit at least once for each anonymous page load, so the lighter and faster it is, the better.